Moved file from root folder to server in order to fix import on `index.ts` (#8693)
* fix import on index.ts and plugin.ts from file on the root of the project.


* changelog


* change import :(


---------
